Split check host env tests
